Do not keep uploaded files in memory, #22
This commit is incomplete because it still keep files
around in the local directory. In the future, we must
add a process responsible for cleaning up temp files
when the process who created them died.

That said, anyone that wants to keep the contents of
the temp file should move it to an appropriate location
otherwise it will be automatically wiped in upcoming
commits.